requirements.txt
requests
tqdm
google-cloud-bigquery
